+++ CHANGELOG.md
@@ -1,9 +1,13 @@
+# 0.0.6 (2021-06-13)
+
+Updated WhiteSpace demo for horizontal mode.
+
 # 0.0.5 (2021-06-12)
 
 Refactor: improve lib mode.
 
 - Updated Flex demo.
-- Updated WhiteSpace demo.
+- Updated WhiteSpace demo for vertical mode.
 
 # 0.0.1 (2021-06-12)